refactor

+++ packages/cubejs-query-orchestrator/src/orchestrator/PreAggregations.ts
@@ -18,7 +18,8 @@ import { cancelCombinator, SaveCancelFn, DriverInterface, BaseDriver,
   DownloadTableData,
   InlineTable,
   StreamOptions,
-  UnloadOptions } from '@cubejs-backend/base-driver';
+  UnloadOptions,
+  DriverCapabilities } from '@cubejs-backend/base-driver';
 import { RedisCacheDriver } from './RedisCacheDriver';
 import { LocalCacheDriver } from './LocalCacheDriver';
 import { Query, QueryCache, QueryTuple, QueryWithParams } from './QueryCache';
@@ -769,13 +770,7 @@ export class PreAggregationLoader {
       if (readOnly) {
         refreshStrategy = this.refreshImplStreamExternalStrategy;
       } else {
-        const capabilities = client?.capabilities?.();
-
-        if (capabilities?.unloadWithoutTempTable && client.unloadWithSql) {
-          refreshStrategy = this.refreshImplWithoutTempTableExternalStrategy;
-        } else {
-          refreshStrategy = this.refreshImplTempTableExternalStrategy;
-        }
+        refreshStrategy = this.refreshImplBucketStrategy;
       }
     }
     return cancelCombinator(
@@ -842,113 +837,53 @@ export class PreAggregationLoader {
     }
   }
 
-  protected async refreshImplWithoutTempTableExternalStrategy(
-    client: DriverInterface,
-    newVersionEntry: VersionEntry,
-    saveCancelFn: SaveCancelFn,
-    invalidationKeys
-  ) {
-    const [sql, params] =
-        Array.isArray(this.preAggregation.sql) ? this.preAggregation.sql : [this.preAggregation.sql, []];
-
-    await client.createSchemaIfNotExists(this.preAggregation.preAggregationsSchema);
-    const targetTableName = this.targetTableName(newVersionEntry);
-
-    const queryOptions = this.queryOptions(invalidationKeys, sql, params, targetTableName, newVersionEntry);
-    this.logExecutingSql(queryOptions);
-
-    try {
-      const tableData = await this.downloadWithoutTempTableExternalPreAggregation(
-        client,
-        newVersionEntry,
-        saveCancelFn,
-        queryOptions,
-        sql,
-        params,
-      );
-
-      try {
-        await this.uploadExternalPreAggregation(tableData, newVersionEntry, saveCancelFn, queryOptions);
-      } finally {
-        if (tableData && tableData.release) {
-          await tableData.release();
-        }
-      }
-    } finally {
-      // We must clean orphaned in any cases: success or exception
-      await this.loadCache.fetchTables(this.preAggregation);
-      await this.dropOrphanedTables(client, targetTableName, saveCancelFn, false, queryOptions);
-    }
-  }
-
-  protected async downloadWithoutTempTableExternalPreAggregation(
-    client: DriverInterface,
+  protected async refreshImplBucketStrategy(
+    client,
     newVersionEntry,
-    saveCancelFn: SaveCancelFn,
-    queryOptions: any,
-    sql: string,
-    sqlParams: unknown[]
+    saveCancelFn,
+    invalidationKeys,
   ) {
-    const table = this.targetTableName(newVersionEntry);
-    this.logger('Downloading external pre-aggregation', queryOptions);
+    const capabilities = client?.capabilities?.();
 
-    try {
-      const externalDriver = await this.externalDriverFactory();
-      const capabilities = externalDriver.capabilities && externalDriver.capabilities();
+    const withTempTable = !(capabilities?.unloadWithoutTempTable && client.unloadWithSql);
 
-      let tableData: DownloadTableData;
-
-      if (capabilities.csvImport && client.unloadWithSql && await client.isUnloadSupported(this.getUnloadOptions())) {
-        tableData = await saveCancelFn(client.unloadWithSql(table, sql, sqlParams, this.getUnloadOptions()));
-      } else {
-        throw new Error('download without temp table error');
-      }
-
-      this.logger('Downloading external pre-aggregation completed', queryOptions);
-
-      return tableData;
-    } catch (error: any) {
-      this.logger('Downloading external pre-aggregation error', {
-        ...queryOptions,
-        error: error?.stack || error?.message
-      });
-      throw error;
-    }
+    return this.runBucketStrategy(
+      client,
+      newVersionEntry,
+      saveCancelFn,
+      invalidationKeys,
+      withTempTable,
+    );
   }
 
   /**
    * Strategy to copy pre-aggregation from source db (with write permissions) to external data
    */
-  protected async refreshImplTempTableExternalStrategy(
+  protected async runBucketStrategy(
     client: DriverInterface,
     newVersionEntry: VersionEntry,
     saveCancelFn: SaveCancelFn,
-    invalidationKeys
+    invalidationKeys,
+    withTempTable: boolean
   ) {
-    const [loadSql, params] =
-        Array.isArray(this.preAggregation.loadSql) ? this.preAggregation.loadSql : [this.preAggregation.loadSql, []];
     await client.createSchemaIfNotExists(this.preAggregation.preAggregationsSchema);
     const targetTableName = this.targetTableName(newVersionEntry);
-    const query = QueryCache.replacePreAggregationTableNames(loadSql, this.preAggregationsTablesToTempTables)
-      .replace(
-        this.preAggregation.tableName,
-        targetTableName
-      );
-    const queryOptions = this.queryOptions(invalidationKeys, query, params, targetTableName, newVersionEntry);
-    this.logExecutingSql(queryOptions);
-    await saveCancelFn(client.loadPreAggregationIntoTable(
+    const queryOptions = this.prepareBucketStrategy(
+      client,
       targetTableName,
-      query,
-      params,
-      queryOptions
-    ));
+      newVersionEntry,
+      saveCancelFn,
+      invalidationKeys,
+      withTempTable,
+    );
 
     try {
-      const tableData = await this.downloadTempExternalPreAggregation(
+      const tableData = await this.downloadExternalPreAggregation(
         client,
         newVersionEntry,
         saveCancelFn,
-        queryOptions
+        queryOptions,
+        withTempTable,
       );
 
       try {
@@ -959,14 +894,69 @@ export class PreAggregationLoader {
         }
       }
     } finally {
+      await this.cleanupBucketStrategy(
+        client,
+        targetTableName,
+        queryOptions,
+        saveCancelFn,
+        withTempTable
+      );
+    }
+  }
+
+  protected async cleanupBucketStrategy(
+    client: DriverInterface,
+    targetTableName: string,
+    queryOptions,
+    saveCancelFn: SaveCancelFn,
+    withTempTable: boolean
+  ) {
+    if (withTempTable) {
       const actualTables = await client.getTablesQuery(this.preAggregation.preAggregationsSchema);
       const mappedActualTables = actualTables.map(t => `${this.preAggregation.preAggregationsSchema}.${t.table_name || t.TABLE_NAME}`);
       if (mappedActualTables.includes(targetTableName)) {
         await client.dropTable(targetTableName);
       }
-      // We must clean orphaned in any cases: success or exception
-      await this.loadCache.fetchTables(this.preAggregation);
-      await this.dropOrphanedTables(client, targetTableName, saveCancelFn, false, queryOptions);
+    }
+
+    // We must clean orphaned in any cases: success or exception
+    await this.loadCache.fetchTables(this.preAggregation);
+    await this.dropOrphanedTables(client, targetTableName, saveCancelFn, false, queryOptions);
+  }
+
+  protected async prepareBucketStrategy(
+    client: DriverInterface,
+    targetTableName: string,
+    newVersionEntry: VersionEntry,
+    saveCancelFn: SaveCancelFn,
+    invalidationKeys,
+    withTempTable: boolean
+  ) {
+    if (withTempTable) {
+      const [loadSql, params] =
+      Array.isArray(this.preAggregation.loadSql) ? this.preAggregation.loadSql : [this.preAggregation.loadSql, []];
+  
+      const query = QueryCache.replacePreAggregationTableNames(loadSql, this.preAggregationsTablesToTempTables)
+        .replace(
+          this.preAggregation.tableName,
+          targetTableName
+        );
+      const queryOptions = this.queryOptions(invalidationKeys, query, params, targetTableName, newVersionEntry);
+      this.logExecutingSql(queryOptions);
+      await saveCancelFn(client.loadPreAggregationIntoTable(
+        targetTableName,
+        query,
+        params,
+        queryOptions
+      ));
+
+      return queryOptions;
+    } else {
+      const [sql, params] =
+        Array.isArray(this.preAggregation.sql) ? this.preAggregation.sql : [this.preAggregation.sql, []];
+      const queryOptions = this.queryOptions(invalidationKeys, sql, params, targetTableName, newVersionEntry);
+      this.logExecutingSql(queryOptions);
+      return queryOptions;
     }
   }
 
@@ -1031,20 +1021,13 @@ export class PreAggregationLoader {
     };
   }
 
-  /**
-   * Create table (for db with write permissions) and extract data via memory/stream/unload
-   */
-  protected async downloadTempExternalPreAggregation(
+  protected async downloadExternalPreAggregation(
     client: DriverInterface,
     newVersionEntry,
     saveCancelFn: SaveCancelFn,
-    queryOptions: any
+    queryOptions: any,
+    withTempTable: boolean
   ) {
-    // @todo Deprecated, BaseDriver already implements it, before remove we need to add check for factoryDriver
-    if (!client.downloadTable) {
-      throw new Error('Can\'t load external pre-aggregation: source driver doesn\'t support downloadTable()');
-    }
-
     const table = this.targetTableName(newVersionEntry);
     this.logger('Downloading external pre-aggregation', queryOptions);
 
@@ -1053,31 +1036,12 @@ export class PreAggregationLoader {
       const capabilities = externalDriver.capabilities && externalDriver.capabilities();
 
       let tableData: DownloadTableData;
-
-      if (capabilities.csvImport && client.unload && await client.isUnloadSupported(this.getUnloadOptions())) {
-        tableData = await saveCancelFn(client.unload(table, this.getUnloadOptions()));
-      } else if (capabilities.streamImport && client.stream) {
-        tableData = await saveCancelFn(
-          client.stream(`SELECT * FROM ${table}`, [], this.getStreamingOptions())
-        );
-
-        if (client.unload) {
-          const stream = new LargeStreamWarning(this.preAggregation.preAggregationId, (msg) => {
-            this.logger('Downloading external pre-aggregation warning', {
-              ...queryOptions,
-              error: msg
-            });
-          });
-          tableData.rowStream.pipe(stream);
-          tableData.rowStream = stream;
-        }
+      if (withTempTable) {
+        tableData = await this.getTableDataWithTempTable(client, table, saveCancelFn, queryOptions, capabilities);
       } else {
-        tableData = await saveCancelFn(client.downloadTable(table, capabilities));
+        tableData = await this.getTableDataWithoutTempTable(client, table, saveCancelFn, queryOptions, capabilities);
       }
 
-      if (!tableData.types) {
-        tableData.types = await saveCancelFn(client.tableColumnTypes(table));
-      }
       this.logger('Downloading external pre-aggregation completed', queryOptions);
 
       return tableData;
@@ -1087,6 +1051,47 @@ export class PreAggregationLoader {
         error: error?.stack || error?.message
       });
       throw error;
+    }
+  }
+
+  protected async getTableDataWithTempTable(client: DriverInterface, table: string, saveCancelFn: SaveCancelFn, queryOptions: any, externalDriverCapabilities: DriverCapabilities) {
+    let tableData: DownloadTableData;
+
+    if (externalDriverCapabilities.csvImport && client.unload && await client.isUnloadSupported(this.getUnloadOptions())) {
+      tableData = await saveCancelFn(client.unload(table, this.getUnloadOptions()));
+    } else if (externalDriverCapabilities.streamImport && client.stream) {
+      tableData = await saveCancelFn(
+        client.stream(`SELECT * FROM ${table}`, [], this.getStreamingOptions())
+      );
+
+      if (client.unload) {
+        const stream = new LargeStreamWarning(this.preAggregation.preAggregationId, (msg) => {
+          this.logger('Downloading external pre-aggregation warning', {
+            ...queryOptions,
+            error: msg
+          });
+        });
+        tableData.rowStream.pipe(stream);
+        tableData.rowStream = stream;
+      }
+    } else {
+      tableData = await saveCancelFn(client.downloadTable(table, externalDriverCapabilities));
+    }
+
+    if (!tableData.types) {
+      tableData.types = await saveCancelFn(client.tableColumnTypes(table));
+    }
+
+    return tableData;
+  }
+
+  protected async getTableDataWithoutTempTable(client: DriverInterface, table: string, saveCancelFn: SaveCancelFn, queryOptions: any, externalDriverCapabilities: DriverCapabilities) {
+    const [sql, sqlParams] =
+        Array.isArray(this.preAggregation.sql) ? this.preAggregation.sql : [this.preAggregation.sql, []];
+    if (externalDriverCapabilities.csvImport && client.unloadWithSql && await client.isUnloadSupported(this.getUnloadOptions())) {
+      return saveCancelFn(client.unloadWithSql(table, sql, sqlParams, this.getUnloadOptions()));
+    } else {
+      throw new Error('download without temp table error');
     }
   }
